enhance precision of internal calculation variables
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2657 d25b4930-3292-e311-b074-0010c6de5da8

+++ libs/Oresme/OresmeKit/OKLineChart.m
@@ -50,13 +50,13 @@
 {
   NSBezierPath *path;
   unsigned i, j;
-  float oneXUnit;
-  float oneYUnit;
-  float xUnitSize;
-  float yUnitSize;
+  double oneXUnit;
+  double oneYUnit;
+  double xUnitSize;
+  double yUnitSize;
   float availableHeight;
   float availableWidth;
-  float rangeToRepresent;
+  double rangeToRepresent;
   float axisLevel;
   float minXPos;
   float minYPos;
@@ -143,7 +143,7 @@
   else if (yAxisGridSizing == OKGridKiloMega)
     {
       int yAxisScaleExp;
-      float scaledRange;
+      double scaledRange;
 
       yAxisScaleExp = 0;
       scaledRange = rangeToRepresent;
